# 校园二手交易APP/openapi.yaml
openapi: 3.0.3
info:
  title: 校园二手交易 APP API
  version: 1.0.0
  description: |
    Android 客户端使用的校园二手交易平台后端接口。所有定义与 `api.md`、`db_schema.md.md`、`spec.md`、
    《后端 API 接口开发规划文档》保持一致。
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.campus-trade.com/api/v1
    description: 生产示例
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: OK
        errorCode:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
        data:
          description: 具体数据载荷
          nullable: true
...
    UserSummary:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        role:
          type: string
          enum: [STUDENT, TEACHER, STAFF]
        email: { type: string }
        phone: { type: string, nullable: true }
        avatar: { type: string, nullable: true }
        school: { type: string, nullable: true }
        emailVerified: { type: boolean }
    Product:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string, nullable: true }
        price: { type: number, format: double }
        category:
          type: string
          enum: [BOOKS, ELECTRONICS, CLOTHING, SPORTS, DAILY, OTHER]
        categoryId:
          type: integer
          nullable: true
          description: 动态分类ID（可选）
        categoryCode:
          type: string
          nullable: true
          description: 动态分类编码（可选）
        categoryName:
          type: string
          nullable: true
          description: 动态分类名称（可选）
        images:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [DRAFT, ON_SALE, OFF_SALE, SOLD, DELETED]
        auditStatus:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        viewCount: { type: integer }
        likeCount: { type: integer }
        contactInfo: { type: string, nullable: true }
        location: { type: string, nullable: true }
        remark: { type: string, nullable: true }
        seller:
          $ref: '#/components/schemas/UserSummary'
        createTime: { type: string, format: date-time }
        favorited: { type: boolean }
...
    Order:
      type: object
      properties:
        id: { type: integer }
        orderNo: { type: string }
        product: { $ref: '#/components/schemas/Product' }
        buyer: { $ref: '#/components/schemas/UserSummary' }
        seller: { $ref: '#/components/schemas/UserSummary' }
        price: { type: number, format: double }
        status:
          type: string
          enum: [PENDING_PAYMENT, PENDING_SHIPMENT, PENDING_RECEIPT, COMPLETED, CANCELLED]
        paymentStatus:
          type: string
          enum: [NOT_INITIATED, PENDING, PROCESSING, SUCCEEDED, FAILED, EXPIRED, CANCELLED, REFUND_PENDING, REFUNDED]
        paymentMethod:
          type: string
          enum: [ALIPAY, WECHAT, CASH]
          nullable: true
        paymentReference: { type: string, nullable: true }
        paymentExpireTime: { type: string, format: date-time, nullable: true }
        paymentTime: { type: string, format: date-time, nullable: true }
        paymentMetadata:
          type: object
          nullable: true
          additionalProperties: true
        deliveryTime: { type: string, format: date-time, nullable: true }
        receiveTime: { type: string, format: date-time, nullable: true }
        buyerNote: { type: string, nullable: true }
        sellerNote: { type: string, nullable: true }
        buyerRating: { type: integer, nullable: true, minimum: 1, maximum: 5 }
        buyerComment: { type: string, nullable: true }
        sellerRating: { type: integer, nullable: true, minimum: 1, maximum: 5 }
        sellerComment: { type: string, nullable: true }
        refundStatus:
          type: string
          enum: [NONE, REQUESTED, PROCESSING, REFUNDED, FAILED]
        refundTime: { type: string, format: date-time, nullable: true }
...
    PaymentIntentResponse:
      type: object
      properties:
        paymentId: { type: integer }
        orderId: { type: integer }
        referenceNo: { type: string }
        status:
          type: string
          enum: [PENDING, PROCESSING, SUCCEEDED, FAILED, EXPIRED, CANCELLED, REFUND_PENDING, REFUNDED]
        method:
          type: string
          enum: [ALIPAY, WECHAT, CASH]
        channel:
          type: string
          enum: [MOCK_QR, MOCK_TRANSFER, CASH]
          nullable: true
        amount: { type: number, format: double }
        currency: { type: string }
        expiresAt: { type: string, format: date-time }
        qrCodeUrl: { type: string, nullable: true }
        payload:
          type: object
          additionalProperties: true
    PaymentRecord:
      type: object
      properties:
        paymentId: { type: integer }
        referenceNo: { type: string }
        status:
          type: string
          enum: [PENDING, PROCESSING, SUCCEEDED, FAILED, EXPIRED, CANCELLED, REFUND_PENDING, REFUNDED]
        method:
          type: string
          enum: [ALIPAY, WECHAT, CASH]
        channel:
          type: string
          enum: [MOCK_QR, MOCK_TRANSFER, CASH]
          nullable: true
        amount: { type: number, format: double }
        currency: { type: string }
        expiresAt: { type: string, format: date-time, nullable: true }
        statusTimeline:
          type: array
          items:
            $ref: '#/components/schemas/PaymentStatusEvent'
        requestPayload:
          type: object
          nullable: true
          additionalProperties: true
        responsePayload:
          type: object
          nullable: true
          additionalProperties: true
        webhookStatus:
          type: string
          enum: [NONE, PENDING, SUCCESS, FAILED]
        callbackAt: { type: string, format: date-time, nullable: true }
        createTime: { type: string, format: date-time }
        updateTime: { type: string, format: date-time }
    PaymentStatusEvent:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, PROCESSING, SUCCEEDED, FAILED, EXPIRED, CANCELLED, REFUND_PENDING, REFUNDED]
        timestamp:
          type: string
          format: date-time
    PaymentStatusResponse:
      type: object
      properties:
        orderId: { type: integer }
        orderStatus:
          type: string
          enum: [PENDING_PAYMENT, PENDING_SHIPMENT, PENDING_RECEIPT, COMPLETED, CANCELLED]
        paymentStatus:
          type: string
          enum: [NOT_INITIATED, PENDING, PROCESSING, SUCCEEDED, FAILED, EXPIRED, CANCELLED, REFUND_PENDING, REFUNDED]
        paymentMethod:
          type: string
          enum: [ALIPAY, WECHAT, CASH]
          nullable: true
        paymentReference: { type: string, nullable: true }
        paymentTime: { type: string, format: date-time, nullable: true }
        paymentExpireTime: { type: string, format: date-time, nullable: true }
        paymentMetadata:
          type: object
          nullable: true
          additionalProperties: true
        refundStatus:
          type: string
          enum: [NONE, REQUESTED, PROCESSING, REFUNDED, FAILED]
        refundTime: { type: string, format: date-time, nullable: true }
        latestIntent:
          $ref: '#/components/schemas/PaymentIntentResponse'
        history:
          type: array
          items:
            $ref: '#/components/schemas/PaymentRecord'
...
paths:
...
  /orders:
    post:
      summary: 创建订单
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId]
              properties:
                productId: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Order'
  /orders/me:
    get:
      summary: 我买到的订单
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [all, pending, confirmed, shipping, completed, cancelled]
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 10 }
      responses:
        '200': { description: 列表, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
  /orders/sold:
    get:
      summary: 我卖出的订单
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [all, pending, confirmed, shipping, completed, cancelled]
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 10 }
      responses:
        '200': { description: 列表, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
  /orders/{id}:
    get:
      summary: 订单详情
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Order' }
  /orders/{id}/confirm:
    post:
      summary: 卖家确认订单
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
  /orders/{id}/reject:
    post:
      summary: 卖家拒绝订单
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
  /orders/{id}/cancel:
    post:
      summary: 买家取消订单
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
  /orders/{id}/complete:
    post:
      summary: 买家确认收货
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
  /orders/{id}/seller-review:
    post:
      summary: 卖家评价买家
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
...
  /payments/orders/{orderId}:
    get:
      summary: 查询订单支付状态
      tags: [支付]
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: integer }
        - in: query
          name: includeHistory
          schema: { type: boolean, default: false }
          description: 是否返回全部支付记录
      responses:
        '200':
          description: 支付状态返回
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaymentStatusResponse'
...